Rails: helptext on how to connect to DB (#82)
I'm still learning my way around Docker and so it took me a little while to work out what you're meant to do once you've applied this template in a Codespace: set your `config/database.yml` and then start your app (and that's it!). Hopefully this helptext makes it a bit more clear for others.