docker transparency workspace bind mounts
Note: If you need to mount folders within the dev container into your own containers using docker-outside-of-docker and you do not want to use Docker in Docker  to avoid a potential performance penalty.

Like In GitHub Codespaces, the workspace folder is **available in the same place on the host as it is in the container,**so you can bind workspace contents as you would normally.