Added tests for Rails and gem installation
This is a Rails container, so checking that Rails is installed makes
sense. The container also installs webdrivers, but it's not clear why.
so I left out that test.

The core issue being solved in this PR is the second test, which
is that installing gems (through gem install or bundler) failed
because of permissions. This test fails in the current version of
the container. If you don't redirect the output, you see the
permission error.