Updated the gem install test
Stopped piping output to /dev/null. This test failed in CI and I'm not
sure why without that output, even though it's passing locally with
these changes. The output is also relatively small - two lines if the
gem installs and an error and short stack trace if it fails.

I also switched the gem to an official GitHub gem. I figure it's more
robust to use one of those than a gem that isn't part of an
organization.