Use vscode user to install gems. (#202)
* Use vscode user to install gems.

Updated the line to install the Rails gem to use the vscode user. This
brings it in line with the optional line to install node packages. Also
updated the example line for additional gems.

* Updated version.

* Added tests for Rails and gem installation

This is a Rails container, so checking that Rails is installed makes
sense. The container also installs webdrivers, but it's not clear why.
so I left out that test.

The core issue being solved in this PR is the second test, which
is that installing gems (through gem install or bundler) failed
because of permissions. This test fails in the current version of
the container. If you don't redirect the output, you see the
permission error.

* Updated the gem install test

Stopped piping output to /dev/null. This test failed in CI and I'm not
sure why without that output, even though it's passing locally with
these changes. The output is also relatively small - two lines if the
gem installs and an error and short stack trace if it fails.

I also switched the gem to an official GitHub gem. I figure it's more
robust to use one of those than a gem that isn't part of an
organization.

* Enabled Actions on my fork to reproduce failing test

* Debugging the failing tests in the GitHub Action

The test script is passing when run locally, but not in GitHub actions.

The command is also working as expected locally inside the container.

* Added explicit checks for write permission.

Explicitly checking the RVM gems directories for the current user's
write permissions in the tests. This is also passing locally within a
container built using the updated configuration and fails within a
container build using the current baseline and in GitHub Actions.

* Trying to understand why default fails

* Inspecting the directories and permissions in gems

* Getting a lot closer to understanding the issue

* Forcing RVM permissions to be fixed.

It doesn't appear that, after installing the gems, users in the rvm
group have write permission to the default directory where these
gems are installed. Running this command fixes that.

* Errors when combining gem install with rvm fix.

* Running RVM permission fix as vscode user

* Explicit path to RVM executable

* Cleaned up tests